Move handlers to separate files.
