+++ src/app_engine/apprtc.py
@@ -26,56 +26,14 @@ import decline_page
 import gcm_register
 import gcmrecord
 import join_page
+import leave_page
+import message_page
 import parameter_handling
 import room as room_module
 import util
 
 jinja_environment = jinja2.Environment(
     loader=jinja2.FileSystemLoader(os.path.dirname(__file__)))
-
-class LeavePage(webapp2.RequestHandler):
-  def post(self, room_id, client_id):
-    result = room_module.remove_client_from_room(
-        self.request.host_url, room_id, client_id)
-    if result['error'] is None:
-      logging.info('Room ' + room_id + ' has state ' + result['room_state'])
-
-class MessagePage(webapp2.RequestHandler):
-  def write_response(self, result):
-    content = json.dumps({ 'result' : result })
-    self.response.write(content)
-
-  def send_message_to_collider(self, room_id, client_id, message):
-    logging.info('Forwarding message to collider for room ' + room_id +
-                 ' client ' + client_id)
-    wss_url, wss_post_url = get_wss_parameters(self.request)
-    url = wss_post_url + '/' + room_id + '/' + client_id
-    result = urlfetch.fetch(url=url,
-                            payload=message,
-                            method=urlfetch.POST)
-    if result.status_code != 200:
-      logging.error(
-          'Failed to send message to collider: %d' % (result.status_code))
-      # TODO(tkchin): better error handling.
-      self.error(500)
-      return
-    self.write_response(constants.RESPONSE_SUCCESS)
-
-  def post(self, room_id, client_id):
-    message_json = self.request.body
-    result = room_module.save_message_from_client(
-        self.request.host_url, room_id, client_id, message_json)
-    if result['error'] is not None:
-      self.write_response(result['error'])
-      return
-    self.write_response(constants.RESPONSE_SUCCESS)
-    if not result['saved']:
-      # Other client joined, forward to collider. Do this outside the lock.
-      # Note: this may fail in local dev server due to not having the right
-      # certificate file locally for SSL validation.
-      # Note: loopback scenario follows this code path.
-      # TODO(tkchin): consider async fetch here.
-      self.send_message_to_collider(room_id, client_id, message_json)
 
 class MainPage(webapp2.RequestHandler):
   def write_response(self, target_page, params={}):
@@ -124,8 +82,8 @@ app = webapp2.WSGIApplication([
     ('/', MainPage),
     ('/decline/(\w+)', decline_page.DeclinePage),
     ('/join/(\w+)', join_page.JoinPage),
-    ('/leave/(\w+)/(\w+)', LeavePage),
-    ('/message/(\w+)/(\w+)', MessagePage),
+    ('/leave/(\w+)/(\w+)', leave_page.LeavePage),
+    ('/message/(\w+)/(\w+)', message_page.MessagePage),
     ('/params', ParamsPage),
     ('/register/(\w+)', gcm_register.BindPage),
     ('/r/(\w+)', RoomPage),
